+++ src/pages/Menu/MenuDetail/MenuDetail.scss
